requirements.txt
Django==5.0.6
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-filter==24.3
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-environ==0.11.2
Pillow==10.4.0
matplotlib==3.8.4

# Dev / Tests
pytest==8.2.1
pytest-django==4.8.0
